feat: add project management components including sidebar, task management, team management, and ideas bank
