+++ client/react-frontend/src/components/demo/ideas-bank.tsx
