+++ client/react-frontend/src/components/demo/info-panel.tsx
