+++ client/react-frontend/src/components/demo/notes-section.tsx
