+++ client/react-frontend/src/components/demo/quick-actions.tsx
