+++ client/react-frontend/src/components/docs/RegistroDeEntrada.tsx
@@ -1,46 +1,46 @@
 import { useEffect, type FC } from "react";
-import { Button } from "./ui/button";
+import { Button } from "../ui/button";
 import { AlertCircle, FileIcon, Plus, SaveAll } from "lucide-react";
 import useEditableTable from "@/hooks/useEditableTable";
-import Table, { Td } from "./BasicTableComponents";
+import Table, { Td } from "../BasicTableComponents";
 import useConnection from "@/hooks/connection/useConnection";
+import useGetEntryRegister from "@/hooks/connection/useGetEntryRegister";
 import { useToast } from "@/hooks/use-toast";
-import useGetExitRegister from "@/hooks/connection/useGetExitRegister";
-import Loading from "./Loading";
+import Loading from "../Loading";
 
-export type ExitRegisterData = {
-	no_exit_register: string;
+export type RegistroDeEntradaData = {
+	numero_registro: string;
 	fecha: string;
-	no_registro_entrada: string;
-	serie_documental: string;
+	tipo_documento: string;
 	sujeto_productor: string;
-	causa_de_salida: string;
+	titulo: string;
+	observaciones?: string;
 };
 
-export type ExitRegisterProps = {
+export type RegistroDeEntradaProps = {
 	archiveId: string;
 	company: string;
 	managementFile: string;
 };
 
-const ExitRegister: FC<ExitRegisterProps> = ({
+const RegistroDeEntrada: FC<RegistroDeEntradaProps> = ({
 	archiveId,
 	company,
 	managementFile,
 }) => {
 	const { rows, addRow, updateRow, saveAllRows, setPrevious } =
-		useEditableTable<ExitRegisterData>([
+		useEditableTable<RegistroDeEntradaData>([
 			{
-				no_exit_register: "",
+				numero_registro: "",
 				fecha: "",
-				no_registro_entrada: "",
-				serie_documental: "",
+				tipo_documento: "",
 				sujeto_productor: "",
-				causa_de_salida: "",
+				titulo: "",
+				observaciones: "",
 			},
 		]);
-	const { handleNewDocumentExit } = useConnection();
-	const { exit, error, loading } = useGetExitRegister(archiveId);
+	const { handleNewDocumentEntry } = useConnection();
+	const { entry, error, loading } = useGetEntryRegister(archiveId);
 	const { toast } = useToast();
 
 	const handleSaveData = (data: string) => {
@@ -48,7 +48,7 @@ const ExitRegister: FC<ExitRegisterProps> = ({
 			alert("No hay datos para guardar");
 			return;
 		}
-		handleNewDocumentExit(
+		handleNewDocumentEntry(
 			data,
 			archiveId,
 			() => {
@@ -69,28 +69,27 @@ const ExitRegister: FC<ExitRegisterProps> = ({
 
 	// biome-ignore lint/correctness/useExhaustiveDependencies: <explanation>
 	useEffect(() => {
-		if (exit && Array.isArray(exit)) {
-			// Combina todos los exit_register de cada objeto en un solo array
-			const allRegisters = exit
-				.filter((item) => Array.isArray(item.exit_register))
-				.flatMap((item) => item.exit_register);
+		if (entry && Array.isArray(entry)) {
+			const allRegisters = entry
+				.filter((item) => Array.isArray(item.entry_register))
+				.flatMap((item) => item.entry_register);
 
 			if (allRegisters.length > 0) {
 				setPrevious(allRegisters);
 			} else {
 				setPrevious([
 					{
-						no_exit_register: "",
+						numero_registro: "",
 						fecha: "",
-						no_registro_entrada: "",
-						serie_documental: "",
+						tipo_documento: "",
 						sujeto_productor: "",
-						causa_de_salida: "",
+						titulo: "",
+						observaciones: "",
 					},
 				]);
 			}
 		}
-	}, [exit, archiveId]);
+	}, [entry, archiveId]);
 
 	if (error) {
 		return (
@@ -131,7 +130,7 @@ const ExitRegister: FC<ExitRegisterProps> = ({
 									<th colSpan={11} className="py-2">
 										<div className="flex">
 											<div className="w-full">ANEXO</div>
-											<div className="w-full text-right"> A4</div>
+											<div className="w-full text-right"> A2</div>
 										</div>
 									</th>
 								</tr>
@@ -140,7 +139,7 @@ const ExitRegister: FC<ExitRegisterProps> = ({
 										colSpan={11}
 										className="text-center py-2 text-base uppercase"
 									>
-										<div>REGISTRO DE SALIDA DE DOCUMENTOS</div>
+										<div>REGISTRO DE ENTRADA DE DOCUMENTOS</div>
 									</th>
 								</tr>
 								<tr>
@@ -162,11 +161,11 @@ const ExitRegister: FC<ExitRegisterProps> = ({
 										// biome-ignore lint/suspicious/noArrayIndexKey: <explanation>
 										<tr key={index}>
 											<Td
-												label={props.no_exit_register}
+												label={props.numero_registro}
 												onBlur={(e) => {
 													updateRow(
 														index,
-														"no_exit_register",
+														"numero_registro",
 														e.currentTarget.innerText,
 													);
 												}}
@@ -178,21 +177,11 @@ const ExitRegister: FC<ExitRegisterProps> = ({
 												}}
 											/>
 											<Td
-												label={props.no_registro_entrada}
+												label={props.tipo_documento}
 												onBlur={(e) => {
 													updateRow(
 														index,
-														"no_registro_entrada",
-														e.currentTarget.innerText,
-													);
-												}}
-											/>
-											<Td
-												label={props.serie_documental}
-												onBlur={(e) => {
-													updateRow(
-														index,
-														"serie_documental",
+														"tipo_documento",
 														e.currentTarget.innerText,
 													);
 												}}
@@ -207,13 +196,18 @@ const ExitRegister: FC<ExitRegisterProps> = ({
 													);
 												}}
 											/>
-
 											<Td
-												label={props.causa_de_salida}
+												label={props.titulo}
+												onBlur={(e) => {
+													updateRow(index, "titulo", e.currentTarget.innerText);
+												}}
+											/>
+											<Td
+												label={props.observaciones ?? ""}
 												onBlur={(e) => {
 													updateRow(
 														index,
-														"causa_de_salida",
+														"observaciones",
 														e.currentTarget.innerText,
 													);
 												}}
@@ -248,29 +242,25 @@ const ExitRegister: FC<ExitRegisterProps> = ({
 	);
 };
 
-export default ExitRegister;
+export default RegistroDeEntrada;
 
 const TableHeading: FC = () => {
 	return (
 		<>
 			<tr className="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-slate-900 dark:text-gray-400">
-				<th className="p-2 border dark:border-gray-700 min-w-20">
-					No. Reg. Salida
+				<th className="p-2 border dark:border-gray-700">No. Reg. Entrada</th>
+				<th className="p-2 border dark:border-gray-700 text-left">Fecha</th>
+				<th className="p-2 border dark:border-gray-700 text-center">
+					Tipo de Documento
 				</th>
-				<th className="p-2 border dark:border-gray-700 text-left min-w-20">
-					Fecha
-				</th>
-				<th className="p-2 border dark:border-gray-700 text-center min-w-20">
-					No. Reg. Entr.
-				</th>
-				<th className="p-2 border dark:border-gray-700 text-center w-full">
-					Serie Documental
-				</th>
-				<th className="p-2 border dark:border-gray-700 text-center min-w-20">
+				<th className="p-2 border dark:border-gray-700 text-center">
 					Sujeto Productor
 				</th>
-				<th className="p-2 border dark:border-gray-700 text-center min-w-20">
-					Causa de Salida
+				<th className="p-2 border dark:border-gray-700 text-center w-full">
+					Título
+				</th>
+				<th className="p-2 border dark:border-gray-700 text-center w-full">
+					Observaciones
 				</th>
 			</tr>
 		</>
