+++ client/react-frontend/src/components/projects/ProjectSidebar.tsx
