+++ client/react-frontend/src/components/projects/top-navigation.tsx
@@ -1,145 +1,138 @@
-"use client"
-
-import { useState } from "react"
-import { ChevronDown, Home, Menu, Settings, User, LogOut, Bell } from "lucide-react"
-import { Button } from "@/components/ui/button"
+"use client";
+import { Button } from "@/components/ui/button";
 import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu"
-import { Badge } from "@/components/ui/badge"
-import { SettingsModal } from "./settings-modal"
+	Select,
+	SelectContent,
+	SelectItem,
+	SelectTrigger,
+	SelectValue,
+} from "@/components/ui/select";
+import useProjects from "@/hooks/connection/useProjects";
+//import useTheme from "@/hooks/useTheme";
+import { Home, ChevronRight } from "lucide-react";
+import UserProfileTrigger from "../UserProfileTrigger";
+import ButtonSwitchTheme from "../ButtonSwitchTheme";
 
 interface TopNavigationProps {
-  selectedProject: string
-  onProjectChange: (projectId: string) => void
-  onHomeClick: () => void
-  onMobileSidebarToggle: () => void
-  isHomePage: boolean
+	selectedProject: string;
+	onProjectChange: (project: string) => void;
+	onHomeClick: () => void;
+	onMobileSidebarToggle: () => void;
+	isHomePage?: boolean;
 }
 
 export function TopNavigation({
-  selectedProject,
-  onProjectChange,
-  onHomeClick,
-  onMobileSidebarToggle,
-  isHomePage,
+	selectedProject,
+	onProjectChange,
+	onHomeClick,
+	onMobileSidebarToggle,
 }: TopNavigationProps) {
-  const [isSettingsOpen, setIsSettingsOpen] = useState(false)
+	const { projects } = useProjects();
 
-  const projects = [
-    { id: "sysgd", name: "SYSGD - Sistema Principal", color: "bg-blue-500" },
-    { id: "marketing", name: "Proyecto Marketing", color: "bg-green-500" },
-    { id: "desarrollo", name: "Desarrollo Web", color: "bg-purple-500" },
-    { id: "investigacion", name: "I+D Innovación", color: "bg-orange-500" },
-  ]
+	const currentProject = projects.find((p) => p.id === selectedProject);
 
-  const currentProject = projects.find((p) => p.id === selectedProject)
+	return (
+		<header className="bg-white dark:bg-gray-900 border-b border-gray-200 dark:border-gray-700 px-4 py-3 sticky top-0 z-50">
+			<div className="flex items-center justify-between">
+				<div className="flex items-center gap-2 md:gap-4">
+					<Button
+						variant="ghost"
+						size="sm"
+						className="md:hidden"
+						onClick={onMobileSidebarToggle}
+					>
+						{/* biome-ignore lint/a11y/noSvgWithoutTitle: <explanation> */}
+						<svg
+							className="w-5 h-5"
+							fill="none"
+							stroke="currentColor"
+							viewBox="0 0 24 24"
+						>
+							<path
+								strokeLinecap="round"
+								strokeLinejoin="round"
+								strokeWidth={2}
+								d="M4 6h16M4 12h16M4 18h16"
+							/>
+						</svg>
+					</Button>
 
-  return (
-    <>
-      <header className="bg-white dark:bg-gray-800 border-b border-gray-200 dark:border-gray-700 px-4 py-3">
-        <div className="flex items-center justify-between">
-          {/* Left Section */}
-          <div className="flex items-center space-x-4">
-            {/* Mobile Menu Button */}
-            {!isHomePage && (
-              <Button variant="ghost" size="sm" className="md:hidden" onClick={onMobileSidebarToggle}>
-                <Menu className="w-5 h-5" />
-              </Button>
-            )}
+					{/* Breadcrumb */}
+					<div className="flex items-center gap-2">
+						<Button
+							variant="ghost"
+							size="sm"
+							onClick={onHomeClick}
+							className={"flex items-center gap-2"}
+						>
+							<Home className="w-4 h-4" />
+							<span className="hidden sm:inline">Inicio</span>
+						</Button>
 
-            {/* Home Button */}
-            <Button variant="ghost" size="sm" onClick={onHomeClick} className="flex items-center space-x-2">
-              <Home className="w-4 h-4" />
-              <span className="hidden sm:inline">Inicio</span>
-            </Button>
+						<>
+							<ChevronRight className="w-4 h-4 text-gray-400" />
+							<span className="text-sm font-medium text-gray-900 dark:text-white">
+								{currentProject?.name || "Proyecto"}
+							</span>
+						</>
+					</div>
 
-            {/* Project Selector */}
-            {!isHomePage && (
-              <DropdownMenu>
-                <DropdownMenuTrigger asChild>
-                  <Button variant="outline" className="flex items-center space-x-2 bg-transparent">
-                    <div className={`w-3 h-3 rounded-full ${currentProject?.color}`} />
-                    <span className="hidden sm:inline max-w-[200px] truncate">{currentProject?.name}</span>
-                    <span className="sm:hidden">Proyecto</span>
-                    <ChevronDown className="w-4 h-4" />
-                  </Button>
-                </DropdownMenuTrigger>
-                <DropdownMenuContent align="start" className="w-64">
-                  {projects.map((project) => (
-                    <DropdownMenuItem
-                      key={project.id}
-                      onClick={() => onProjectChange(project.id)}
-                      className="flex items-center space-x-3"
-                    >
-                      <div className={`w-3 h-3 rounded-full ${project.color}`} />
-                      <span>{project.name}</span>
-                      {project.id === selectedProject && (
-                        <Badge variant="secondary" className="ml-auto">
-                          Activo
-                        </Badge>
-                      )}
-                    </DropdownMenuItem>
-                  ))}
-                </DropdownMenuContent>
-              </DropdownMenu>
-            )}
-          </div>
+					<div className="h-6 w-px bg-gray-300 dark:bg-gray-600 hidden sm:block" />
 
-          {/* Right Section */}
-          <div className="flex items-center space-x-2">
-            {/* Notifications */}
-            <Button variant="ghost" size="sm" className="relative">
-              <Bell className="w-4 h-4" />
-              <Badge className="absolute -top-1 -right-1 w-5 h-5 p-0 flex items-center justify-center text-xs bg-red-500">
-                3
-              </Badge>
-            </Button>
+					<div className="flex items-center gap-2">
+						<div className="w-6 h-6 bg-blue-600 rounded flex items-center justify-center">
+							<span className="text-white text-xs font-bold">S</span>
+						</div>
+						<span className="font-bold text-lg text-gray-900 dark:text-white hidden sm:inline">
+							SYSGD
+						</span>
+					</div>
 
-            {/* Settings Button */}
-            <Button
-              variant="ghost"
-              size="sm"
-              onClick={() => setIsSettingsOpen(true)}
-              className="flex items-center space-x-2"
-            >
-              <Settings className="w-4 h-4" />
-              <span className="hidden sm:inline">Ajustes</span>
-            </Button>
+					<div className="hidden lg:block">
+						<Select value={selectedProject} onValueChange={onProjectChange}>
+							<SelectTrigger className="w-64">
+								<SelectValue placeholder="Seleccionar proyecto" />
+							</SelectTrigger>
+							<SelectContent>
+								{projects.map((project) => (
+									<SelectItem key={project.id} value={project.id}>
+										{project.name}
+									</SelectItem>
+								))}
+							</SelectContent>
+						</Select>
+					</div>
+				</div>
 
-            {/* User Menu */}
-            <DropdownMenu>
-              <DropdownMenuTrigger asChild>
-                <Button variant="ghost" size="sm" className="flex items-center space-x-2">
-                  <div className="w-8 h-8 bg-blue-500 rounded-full flex items-center justify-center text-white text-sm font-medium">
-                    LY
-                  </div>
-                  <span className="hidden md:inline">Lazaro Yunier</span>
-                  <ChevronDown className="w-4 h-4" />
-                </Button>
-              </DropdownMenuTrigger>
-              <DropdownMenuContent align="end" className="w-56">
-                <DropdownMenuItem className="flex items-center space-x-2">
-                  <User className="w-4 h-4" />
-                  <span>Mi Perfil</span>
-                </DropdownMenuItem>
-                <DropdownMenuSeparator />
-                <DropdownMenuItem className="flex items-center space-x-2 text-red-600">
-                  <LogOut className="w-4 h-4" />
-                  <span>Cerrar Sesión</span>
-                </DropdownMenuItem>
-              </DropdownMenuContent>
-            </DropdownMenu>
-          </div>
-        </div>
-      </header>
+				<div className="flex items-center gap-1 md:gap-2">
+					{/* <Button variant="ghost" size="sm" className="hidden sm:flex">
+						<Bell className="w-4 h-4" />
+					</Button>
 
-      {/* Settings Modal */}
-      <SettingsModal isOpen={isSettingsOpen} onClose={() => setIsSettingsOpen(false)} />
-    </>
-  )
+					<Button variant="ghost" size="sm" className="hidden sm:flex">
+						<Settings className="w-4 h-4" />
+					</Button> */}
+					<ButtonSwitchTheme />
+					<div className="flex items-center gap-2 ml-2 pl-2 border-l border-gray-300 dark:border-gray-600">
+						<UserProfileTrigger />
+					</div>
+				</div>
+			</div>
+
+			<div className="lg:hidden mt-3">
+				<Select value={selectedProject} onValueChange={onProjectChange}>
+					<SelectTrigger className="w-full">
+						<SelectValue placeholder="Seleccionar proyecto" />
+					</SelectTrigger>
+					<SelectContent>
+						{projects.map((project) => (
+							<SelectItem key={project.id} value={project.id}>
+								{project.name}
+							</SelectItem>
+						))}
+					</SelectContent>
+				</Select>
+			</div>
+		</header>
+	);
 }
